完成根据id修改用户信息 根据id查询用户 增加了自定义异常 、全区异常处理 枚举类 结果集封装方法

+++ FastView-Backend/src/main/java/cn/kshost/fastview/backend/pojo/po/Role.java
@@ -27,6 +27,7 @@ public class Role implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
+
     /**
      * 主键ID
      */
@@ -61,11 +62,14 @@ public class Role implements Serializable {
      * 状态：1 正常，0 禁用
      */
     @TableField("status")
-    private Byte status;
+    private Integer status;
 
     /**
      * 逻辑删除：0 未删除，1 已删除
      */
     @TableField("is_delete")
-    private Byte isDelete;
+    private Integer isDelete;
+
+
+
 }
